+++ docker/prod/api/rollback.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
 set -e
 
-COMPOSE_FILE="docker-compose-prod.yml"
-HEALTH_TIMEOUT=60
+SCRIPT_DIR="$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" && pwd)"
+COMPOSE_FILE="$SCRIPT_DIR/docker-compose-prod.yml"
+HEALTH_TIMEOUT=120
 HEALTH_INTERVAL=5
 
 log_info() { echo "[INFO] $1"; }
